scripts/gsettings.sh: Make alt-tab behave.
The default Alt-Tab behaviour of Gnome 3 is outright stupidly
idiotically hostile.  If I put a thing on another workspace, that's
because I do not want it on my way at the moment.  If Alt-Tab will
cycle through all of them, what's the use of the workspace?  And if
Alt-` is there, and Super+Tab switches between apps, why not keep
Alt+Tab as "switch between windows in the current workspace"?  Them
guys should quit writing code on drugs.