debian: Stop texlive-full downloading the fucking world.
TeX documentation is in PDF, and texlive-full depends on documentation
packages.  In this patch I include its deps directly into
deb/DEBIAN/deps.mk, stripping -doc packages (and some other stuff),
and add --no-install-recommends flag to the ‘apt-get install’ command
in the deb-inst rule in makefile, so that none of that crap gets
installed.