Add system/debian/etc/alternatives/x-cursor-theme.
Disables Adwaita on Debian.  If you have any GTK apps, Adwaita is a
hard dep (which is kinda understandable), and if you have to have
Adwaita, you have to have its cursors as default too because Adwaita
likes so. update-alternatives(1) does not let me configure this, so I
just write over the /etc/alternatives/x-cursor-theme file with an
empty one, and X default cursors are used.