emacs.d/gk.org: Use run-with-idle-timer in After-save hook to make it less annoying.
The hook used to query as soon as the file was saved.  I have the
tendency to make minor edits and save very frequently (e.g. type: suck
C-x C-s M-b C-d d C-x C-s M-c C-x C-s), and the frequent y-or-n-p
blocking me was annoying.  Now in the new version, it uses an idle
timer to only ask when Emacs is idle for two secods, and if I save a
file before the queries are made, it only prompts me once.