Replace the packages.list with the new layers system
The layers system is baked and now can work through the ‘make
alpha-debian-init’ interface.  This patch involves some fixes to other
parts of the build, specifically the ‘cd x; make; cd ..’ pattern is
replaced w/ the more correct ‘make -C x ...’ pattern.

Also, a Dockerfile is added in order to reliably test the installation
system.  A new make rule ‘alpha-debian-test’ was added for this
purpose.  .gitlab-ci.yml is also updated to use the same image and the
script as this testing setup.