Enable synthing again, only for ~/syn
Syncthing is not safe to use for important data [1], but it is better
than competition b/c does not require a LAMP stack to run.  So enable
it for one folder again to be able to conveniently sync certain files
between devices.

This sync folder should not be treated as a safe space for files.

[1] https://www.gkayaalp.com/blog/20190102_syncthing.html