emacs.d/init.el: open projects in new frame
Open a project in a new frame, which means, given a path to a project
directory, open a new frame, open that path in dired and magit/vc,
start a shell for that project, put that shell’s buffer name into the
new frames ‘gk-project-shell’ parameter.

Use that parameter in ‘gk-pop-shell’ to toggle the project shell if
available.

Bind new function ‘gk-open-project’ to gk-prefix + C-p.