Switch to Debian GNOME
This includes the removal of quite a bit of configuration,
including all of xdg-config/.  Dconf dumps are frequently taken
and saved to lib/ instead.

Packages this edition of Debian installs by default are removed from
the package list.  User is no longer created by the initialisation
scripts.  No XFCE4 apps are used.