Move my packages to a separate module, add gk-mairix, upgrade qutebrowser to v2.0.2
This patch also includes necessary environment setup to integrate this
with Guile and Guix too.  The way the relevant path variables need to
be configured is somewhat obscure.  In order for both Guile and Guix
to work as expected, a configuration like follows is necessary:

export GUILE_LOAD_PATH=$MYLIB/scheme:$GUILE_LOAD_PATH
export GUIX_PACKAGE_PATH=$MYLIB/scheme

If the Guile path is not set correctly, a package path like (gk
guix-packages) cannot be used with Guix, and the second variable would
look like this:

export GUIX_PACKAGE_PATH=$MYLIB/scheme/gk

Furthermore, you wouldn’t be able to define the module as (gk
guix-packages), but instead it would have to be just (guix-packages).

Lastly, this patch adds a package definition for mairix.