De-systemd-ify user configuration
User level systemd is a PITA to work with, and is not portable to
GuixSD, other non-systemd distros, and BSDs.

So user daemons are moved into ‘lib/launch-session-servers.bash’ for
now.  I can at least get them to output logs to somewhere, and find
out why they don’t work.