emacs.d/init.el: move theme customisations into a hook
Theme customisations were all part of ‘gk-setup-frame-looks’, which
made that a Long Ass Function 🐴™.

This patch moves that stuff outside of the function, now a hook called
‘gk-gui-theme-customisation-functions’ is used to deal with that.

The idea is, each function checks it’s single argument to see if it’s
a theme it wants to customise, and if yes, does the customisation.  We
don’t care if they do anything or returns any value or how many of
them match and run.