lib/prompt.bash: speed up and re-enable VCS status in prompt
Manually check for supported repositories instead of using the VCS tool
commands.  Edge cases may be unsupported (like a repo with some worktree
setting where the .git is not near the files) but for most cases should
work.  Basically a simple upwards traversal of the path components of
the current working directory and looking for known VCS directories.

Probably wouldn't support Fossil repositories.