systems/sappho/desktop-setup.bash: do not stack shadows, do not shadow i3 window titles
Before these settings the window titles (which showed up in stacked
and tabbed layouts in i3-gaps) would get shadowns, and in the
mentioned layouts, the shadows of the stacked windows would get
stacked themselves, resulting in lopsided and heavy shadows.

Now we ask picom to avoid shadowing those titles, and to only shadow
the active window.

This command line is getting ugly, maybe I need to actually give the
thingy a config file...