bin/batmon.bash: filter uevent file for whitespace before loading
Some rhs values can have spaces in them, which breaks the script when
the file is sourced. Filter the file to convert horizontal whitespace
to nonspace characters so that the script is robust.

This started happening after an update so I suppose they just changed
the way the uevent file is written recently.