dotfiles/profile.sh: simplify how $ENV is set and loaded
Simply set $ENV to the POSIX compatible shrc instead of trying to
determine the current shell. That shrc should do the job of
determining the shell and loading the specific RC file.

This will require change if I ever use a non-POSIX-compatible shell as
my default shell for my user. Unlikely...