Don’t log persistently
Session logs can be verbose and don’t need to persist, so moved logs
to /tmp, and stopped logging from msmtp (procmail logs should suffice,
msmtprc does not seem to handle environment variables in it’s command
lines).

This also means logs are no longer backed up. They instead exist for
the duration of the session in $MYLOGS/.

This should reduce the load on the storage unit and reduce the number
of failed borg backups due to xsession log in particular changing
while backup is in progress.

(Xsession log is particularly verbose due to how verbose i3’s logging
is.)