lib/prompt.bash: fix prompt clobbered, add simple prompt
Turns out what clobbered the prompt line was bp_lastcmdexit,
specifically, its output being on the same line as the cursor. I suppose
multiple ``tput sgr0'' on the same line break something. With this
commit, the error message gets its own line.

Furthermore, a simple prompt is added, all the function local variables
are marked as such, and also the reprinting of the erroneous exit message
is suppressed by bp_lastcmdexit tracking and checking on HISTCMD.